2_pc_datahandler/pyproject.toml
[build-system]
requires = ["hatchling", "wheel"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.sdist]
where = ["."]
include = ["eeg_api"]
exclude = ['*_test.py', 'dist', 'example', 'test_data']
namespaces = false


[tool.hatch.build.targets.wheel]
packages = ["eeg_api"]
exclude = ['*_test.py', 'dist', 'example', 'test_data']
only-packages = true


[project]
name = "4_pc_datahandler"
version = "0.1.0"
description = "Python API for processing recorded EEG data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pylsl>=1.17.6",
    "pyqt5>=5.15.11",
    "pyqtgraph>=0.14.0",
    "pyserial>=3.5",
]


[tool.uv]
default-groups = ["dev", "analysis"]
required-environments = [
  "sys_platform == 'win32' and platform_machine == 'AMD64'",
]


[dependency-groups]
dev = [
    "hatch>=1.14.0",
    "hatchling>=1.27.0",
    "wheel>=0.45.1",
]
analysis = [
    "elasticai.hw_measurements @ git+https://github.com/es-ude/elastic-ai.hw-measurements.git",
]
